feat: 디테일페이지 이름 변경 및 리팩토링 좀 더 추가
