+++ src/hooks/detailpage/useSessionData.ts
@@ -1,0 +1,9 @@
+import { fetchSessionData } from "@/utils/fetchSession";
+import { useQuery } from "@tanstack/react-query";
+
+export const useSessionData = () => {
+  return useQuery({
+    queryKey: ["session"],
+    queryFn: fetchSessionData,
+  });
+};
